Fix compilation errors and warnings
1. Fixed right shift overflow warning in circular_buffer.hpp
   - Added SIZE_MAX check to only perform 32-bit right shift on 64-bit systems
   - Optimized nextPowerOfTwo function implementation

2. Fixed ESP-IDF I2C driver formatting error
   - Changed %u format specifier to %lu with type casting
   - Added inttypes.h and driver/i2c.h header files

3. Resolved I2C driver RTTI issue
   - Replaced dynamic_pointer_cast with static_pointer_cast
   - Removed noStop parameter from writeReadDevice function

修复编译错误和警告

1. 修复 circular_buffer.hpp 中的右移位溢出警告
   - 添加 SIZE_MAX 检查，仅在 64 位系统上执行 32 位右移
   - 优化 nextPowerOfTwo 函数实现

2. 修复 ESP-IDF I2C 驱动格式化错误
   - 将 %u 格式说明符改为 %lu 并添加类型转换
   - 添加 inttypes.h 和 driver/i2c.h 头文件

3. 解决 I2C 驱动 RTTI 问题
   - 将 dynamic_pointer_cast 替换为 static_pointer_cast
   - 移除 writeReadDevice 函数中的 noStop 参数

+++ src/device/bsp/esp32_s3/laminpie_esp_i2c_bus.cpp
@@ -3,6 +3,8 @@
 #include "esp_log.h"
 #include <cstdint>
 #include <sys/_stdint.h>
+#include <inttypes.h>
+#include "driver/i2c.h"
 
 EspI2cBus::EspI2cBus(i2c_port_num_t busNumber, gpio_num_t sda_pin, gpio_num_t scl_pin, uint32_t frequency)
     : I2c_bus(busNumber, sda_pin, scl_pin, frequency),
@@ -21,7 +23,7 @@ EspI2cBus::~EspI2cBus() {
 }
 
 bool EspI2cBus::initialize() {
-    ESP_LOGI(TAG, "Initializing I2C on SDA: %d, SCL: %d with frequency: %d Hz", sda_pin_, scl_pin_, frequency_);
+    ESP_LOGI(TAG, "Initializing I2C on SDA: %d, SCL: %d with frequency: %lu Hz", sda_pin_, scl_pin_, (unsigned long)frequency_);
 
     esp_err_t ret;
         i2c_master_bus_config_t i2c_bus_cfg = {
